重构 把question的score改为double 添加 QuestionInfo添加分数

+++ app/src/main/java/com/edu_app/model/teacher/practice/QuestionItem.java
@@ -70,11 +70,11 @@ public class QuestionItem implements Model, Serializable {
         return question.getQuestionType();
     }
 
-    public int getScore(){
+    public double getScore(){
         return question.getScore();
     }
 
-    public void setScore(int score){
+    public void setScore(double score){
         question.setScore(score);
     }
 }
